Move even more into Shared
